Split macro in separate module
