+++ red4ext-sys/src/interop.rs
@@ -419,20 +419,6 @@ impl ResourcePathBuilder {
     }
 }
 
-/// shortcut for ResourcePath creation.
-#[macro_export]
-macro_rules! res_path {
-    ($base:expr, /$lit:literal $($tt:tt)*) => {
-        $crate::res_path!($base.join($lit), $($tt)*)
-    };
-    ($base:expr, ) => {
-        $base
-    };
-    ($lit:literal $($tt:tt)*) => {
-        $crate::res_path!($crate::interop::ResourcePath::builder().join($lit), $($tt)*).try_build()
-    };
-}
-
 #[cfg(test)]
 mod tests {
     use std::path::PathBuf;
@@ -499,15 +485,5 @@ mod tests {
                 hash: fnv1a64("multi\\somewhere\\in\\archive")
             }
         );
-    }
-    #[cfg(feature = "macros")]
-    #[test]
-    fn res_path() {
-        use crate::res_path;
-        assert!(res_path!("").is_err());
-        assert!(res_path!(".." / "somewhere" / "in" / "archive" / "custom.ent").is_err());
-        assert!(res_path!("base" / "somewhere" / "in" / "archive" / "custom.ent").is_ok());
-        assert!(res_path!("custom.ent").is_ok());
-        assert!(res_path!(".custom.ent").is_ok());
     }
 }
