+++ red4ext-sys/src/lib.rs
@@ -1,5 +1,7 @@
 pub mod error;
 pub mod interop;
+#[cfg(feature = "macros")]
+pub mod macros;
 
 #[allow(clippy::missing_safety_doc)]
 #[cxx::bridge]
